Users can be deleted * projects are shown as removed vs not defined. * confirm before delete project and student. * delete button decrements the project associated with it.

+++ app/routes/signup/organization/edit.js
@@ -23,6 +23,8 @@ export default Ember.Route.extend({
     },
     destroy() {
       var self = this;
+      if(confirm("Are you sure you want to delete this project?") != true)
+        return;
       this.modelFor(this.routeName).destroyRecord().then(function() {
         self.transitionTo('signup.organization');
       });
